Make replay playback speed affect logic framerate instead of skipping frames

+++ src/replay.h
@@ -39,7 +39,7 @@ struct replay {
 	struct replayHeader header;
 	queue *frames;
 	bool playback;
-	double frame;
+	int frame;
 	int totalFrames;
 	float speed;
 };
